Ensure drained nodes are uncordoned if test times out
In the pod recreation test, when deployments are not restarted in time
after a node is drained, that node would remain drained forever and
never be uncordoned.

We can defer the node uncordon operation, so it will always be executed.
If the node was successfully uncordoned, a second uncordon will become a
no-op.

We also remove some unneeded checks after the call to uncordonNode(),
since they are already performed by waitForAllDeploymentsReady(), and
move the final waitForAllDeploymentsReady() check outside of the loop,
so we only do it once.